requirements.txt
async-timeout==5.0.1
beautifulsoup4==4.13.5
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
Flask==3.1.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
lxml==6.0.1
lyricsgenius==3.0.1
MarkupSafe==3.0.2
nltk==3.9.1
python-dotenv==1.1.1
redis==6.4.0
regex==2025.9.18
requests==2.32.5
soupsieve==2.8
spotipy==2.25.1
textblob==0.19.0
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
gunicorn==22.0.0
